% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fit_dlm}
\alias{fit_dlm}
\title{Fit a Bayesian multivariate dynamic linear model with Stan}
\usage{
fit_dlm(
  formula = NULL,
  time_varying = NULL,
  time = "year",
  est_df = FALSE,
  family = c("normal", "binomial", "poisson", "nbinom2", "gamma", "lognormal"),
  correlated_rw = FALSE,
  data,
  chains = 3,
  iter = 2000,
  warmup = floor(iter/2),
  ...
)
}
\arguments{
\item{formula}{The model formula for the fixed effects; at least this formula or \code{time_varying} needs to have the response included}

\item{time_varying}{The model formula for the time-varying effects; at least this formula or \code{formula} needs to have the response included}

\item{time}{String describing the name of the variable corresponding to time, defaults to "year"}

\item{est_df}{Whether or not to estimate deviations of B as Student - t with estimated degrees of freedom, defaults to \code{FALSE}}

\item{family, }{The name of the family used for the response; can be one of "normal","binomial","possion","nbinom2","gamma","lognormal"}

\item{correlated_rw, }{Whether to estimate time-varying parameters as correlated random walk, defaults to FALSE}

\item{data}{The data frame including response and covariates for all model components}

\item{chains}{Number of mcmc chains, defaults to 3}

\item{iter}{Number of mcmc iterations, defaults to 2000}

\item{warmup}{Number iterations for mcmc warmup, defaults to 1/2 of the iterations}

\item{...}{Any other arguments to pass to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}.}
}
\value{
A list containing the fitted model and arguments and data used
to fit the model. These include \code{model} (the fitted model object of class \code{stanfit}),
}
\description{
Fit a Bayesian multivariate dynamic linear model with Stan that optionally includes covariates to estimate
effects, extremes (Student-t distribution), etc.
}
\examples{
\donttest{
set.seed(123)
N = 20
data = data.frame("y" = runif(N),
                  "cov1" = rnorm(N),
                  "cov2" = rnorm(N),
                  "year" = 1:N,
                  "season" = sample(c("A","B"), size=N, replace=TRUE))
b_1 = cumsum(rnorm(N))
b_2 = cumsum(rnorm(N))
data$y = data$cov1*b_1 + data$cov2*b_2
time_varying = y ~ cov1 + cov2
formula = NULL

# fit a model with a time varying component
fit <- fit_dlm(formula = formula,
               time_varying = time_varying,
               time = "year",
               est_df = FALSE,
               family = c("normal"),
               data=data, chains = 1, iter = 20)

# fit a model with a time varying and fixed component (here, fixed intercept)
fit <- fit_dlm(formula = y ~ 1,
               time_varying = y ~ -1 + cov1 + cov2,
               time = "year",
               est_df = FALSE,
               family = c("normal"),
               data=data, chains = 1, iter = 20)

#' # fit a model with deviations modeled with a multivariate Student-t
fit <- fit_dlm(formula = y ~ 1,
               time_varying = y ~ -1 + cov1 + cov2,
               time = "year",
               est_df = TRUE,
               family = c("normal"),
               data=data, chains = 1, iter = 20)

#' #' # fit a model with deviations modeled with a multivariate Student-t
fit <- fit_dlm(formula = y ~ 1,
               time_varying = y ~ -1 + cov1 + cov2,
               time = "year",
               est_df = TRUE,
               family = c("normal"),
               data=data, chains = 1, iter = 20)
}

}
