% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvdlm-package.R
\docType{package}
\name{mvdlm-package}
\alias{mvdlm-package}
\alias{mvdlm}
\title{The 'mvdlm' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
