% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvdlm-package.R
\docType{package}
\name{mvdlm-package}
\alias{mvdlm-package}
\alias{mvdlm}
\title{The 'mvdlm' package.}
\description{
Multivariate dynamic linear models fit with Stan
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/atsa-es/mvdlm}
  \item Report bugs at \url{https://github.com/atsa-es/mvdlm/issues}
}

}
\author{
\strong{Maintainer}: Eric J. Ward \email{eric.ward@noaa.gov} (\href{https://orcid.org/0000-0002-4359-0296}{ORCID})

}
